pyproject.toml
[project]
name = "ibm-skillsbuild"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain-core>=1.2.5",
    "langchain-google-genai>=4.1.2",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "plotly>=6.5.0",
    "pydantic>=2.12.5",
    "pymupdf>=1.26.7",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-pptx>=1.0.2",
    "streamlit>=1.52.2",
]
